backend/requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette==0.27.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis==5.0.1
databases[postgresql]==0.8.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-oauth2==1.1.1
cryptography==41.0.7
bcrypt==4.1.2

# Production Middleware & Monitoring
prometheus-fastapi-instrumentator==6.1.0
email-validator==2.1.0
python-json-logger==2.0.7
structlog==23.2.0

# AI & ML
openai==1.3.7
pinecone-client==2.2.4
azure-cognitiveservices-language-textanalytics==5.3.0
azure-ai-textanalytics==5.3.0

# HTTP & Networking
httpx==0.25.2
requests==2.31.0
websockets==12.0
socketio==0.2.1
python-socketio==5.10.0
aiohttp==3.9.1

# File Processing
python-docx==1.1.0
PyPDF2==3.0.1
python-magic==0.4.27
Pillow==10.1.0

# Email & Communication
sendgrid==6.10.0
emails==0.6.1

# Background Tasks
celery==5.3.4
flower==2.0.1

# Monitoring & Logging
sentry-sdk[fastapi]==1.38.0
loguru==0.7.2

# Data Processing
pandas==2.1.3
numpy==1.25.2
python-dateutil==2.8.2

# Web Scraping (for Corporate DNA)
beautifulsoup4==4.12.2
scrapy==2.11.0
selenium==4.15.2

# Payment Processing
stripe==7.7.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
factory-boy==3.3.0
faker==20.1.0

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Validation & Serialization
marshmallow==3.20.1
pydantic[email]==2.5.0

# Caching
aiocache==0.12.2

# Rate Limiting
slowapi==0.1.9

# CORS
fastapi-cors==0.0.6

# Performance
orjson==3.9.10

# WebRTC Support
aiortc==1.6.0

# Video/Audio Processing
opencv-python==4.8.1.78
librosa==0.10.1

# Cloud Storage & Services
boto3==1.34.0
azure-storage-blob==12.19.0
azure-identity==1.15.0

# Production Utilities
gunicorn==21.2.0
supervisor==4.2.5
python-magic-bin==0.4.14  # For Windows compatibility

# Data Validation & Security
bleach==6.1.0
html5lib==1.1
defusedxml==0.7.1

# Time & Date Utilities
pytz==2023.3
arrow==1.3.0

# Async Utilities
asyncio-mqtt==0.16.2
aiofiles==23.2.1

# Metrics & APM
psutil==5.9.6
py-cpuinfo==9.0.0

# Task Queue Redis
rq==1.15.1
rq-dashboard==0.7.0

# Load Testing
locust==2.17.0

# Azure Services
azure-functions==1.18.0
azure-cosmos==4.5.1
azure-servicebus==7.11.4
azure-keyvault-secrets==4.7.0

# Notification Services
python-pushover==0.4
python-telegram-bot==20.7

# Compression
zstandard==0.22.0
lz4==4.3.2

# SQLite for Development
aiosqlite==0.19.0

# Additional Security
argon2-cffi==23.1.0
python-gnupg==0.5.2